+++ src/rbe510.hpp
@@ -109,8 +109,8 @@ public:
 	void disableVerbose();
 	void drawGrid(int gridSpacing);
 	void hideGrid();
-	cv::Point2i pixelToGridCoordinates(cv::Point2f goal);
-	cv::Point2i gridToPixelCoordinates(cv::Point2f goal);
+	cv::Point2f pixelToGridCoordinates(cv::Point2f goal);
+	cv::Point2f gridToPixelCoordinates(cv::Point2f goal);
 };
